feat(web-app-password-protected-folders): setup app
Add a new web extension called Password Protected Folders. This extension allows users to create a new folder which will be (as the name suggests) protected by a password. This PR sets up the app and adds basic flow to create the folder. In the following PRs, additional features like opening the folder or enforcing password policies will be added.